requirements.txt
click>=8.1
rich>=13.0
prompt_toolkit>=3.0
gitpython>=3.1
litellm>=1.30
tiktoken>=0.6
requests>=2.31
beautifulsoup4>=4.12
watchdog>=4.0
pyyaml>=6.0
pathspec>=0.12
pygments>=2.17
python-dotenv>=1.0
pyperclip>=1.8

# dev
pytest>=7
pytest-cov
black>=23
ruff>=0.1
mypy>=1.0
